DELETED 표식으로 변경, 파일 구조 변경
