+++ src/view/hashtable.html
@@ -161,7 +161,7 @@
     <script src="js/static/jquery-3.3.1.slim.min.js"></script>
     <script src="js/static/popper.min.js"></script>
     <script src="js/static/bootstrap.min.js"></script>
-    <script src="js/hash/index.js"></script>
+    <script src="js/hashtable/index.js"></script>
 
 </body>
